코드 수정 및 데이터 파일 업데이트

+++ src/com/mycom/word/WordCRUD.java
@@ -115,7 +115,7 @@ public class WordCRUD implements ICRUD{
         String ans = s.next();
         if(ans.equalsIgnoreCase("y"))
         {
-            list.remove(idlist.get(id-1));
+            list.remove((int)idlist.get(id-1));
             System.out.println("단어가 삭제되었습니다. ");
         }
         else
